Move service_client_mock to pilz_testutils
